Metadata File moved to core.components
